intermediate commit

+++ Backend-part/ClubberTv.Infrastructure/Migrations/AppDbContextModelSnapshot.cs
@@ -39,8 +39,8 @@ namespace ClubberTV.Infrastructure.Migrations
                     b.Property<int>("Duration")
                         .HasColumnType("int");
 
-                    b.Property<byte>("Status")
-                        .HasColumnType("tinyint");
+                    b.Property<int>("Status")
+                        .HasColumnType("int");
 
                     b.Property<string>("Title")
                         .IsRequired()
